# api/openapi.yaml
openapi: 3.0.3
info:
  title: Toskar Control Plane API
  version: 1.4.0
  description: >-
    Local AI control plane API for Toskar, its OpenAI-compatible routes, and its MCP server.
    On the default loopback bind no key is needed. When the daemon listens on any other address,
    every /api/v1, /v1, and /mcp route needs Authorization: Bearer YOUR_API_KEY; /about and /source do not.
    Errors are {"error": {"code", "message", "details"}}. docs/api.md describes the behavior of each route.
servers:
  - url: http://127.0.0.1:7331
security:
  - {}
  - bearerAuth: []
paths:
...
  /api/v1/profiles:
    get:
      summary: List AI profiles
      operationId: listProfiles
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AIProfile"
    post:
      summary: Create a profile
      operationId: createProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AIProfile"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIProfile"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/api-keys/{id}/permissions:
    put:
      summary: Change what an API key may ask of the assistant
      operationId: setAPIKeyPermissions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                memory:
                  type: string
                  enum: [never, on_request, always]
                knowledge:
                  type: string
                  enum: [never, on_request, always]
                tools:
                  type: string
                  enum: [profile, read_only, none]
                placement:
                  type: boolean
      responses:
        "200":
          description: The key, with its permissions
        "400":
          description: Invalid permissions or unknown key
...
  /api/v1/diagnostics:
    get:
      summary: Create a diagnostics bundle (a zip without secrets or private keys)
      operationId: exportDiagnostics
      parameters:
        - name: include_conversations
          in: query
          schema:
            type: string
            enum: ["true", "false"]
      responses:
        "200":
          description: Where the bundle was written
          content:
            application/json:
              schema:
                type: object
                properties:
                  path:
                    type: string
                  message:
                    type: string
        "500":
          description: The bundle could not be created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a diagnostics bundle
      operationId: createDiagnostics
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                include_conversations:
                  type: boolean
      responses:
        "200":
          description: Where the bundle was written
          content:
            application/json:
              schema:
                type: object
                properties:
                  path:
                    type: string
                  message:
                    type: string
        "500":
          description: The bundle could not be created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/models/recommend:
    get:
      summary: Recommended models for this computer
      operationId: recommendModels
      parameters:
        - name: purpose
          in: query
          description: What the model is for, such as chat or coding
          schema:
            type: string
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
  /api/v1/models/browse:
    get:
      summary: Search Hugging Face for GGUF models
      operationId: browseModels
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Models
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
...
  /api/v1/models/install-from-url:
    post:
      summary: Download a GGUF from a URL
      operationId: installModelFromURL
      parameters:
        - name: wait
          in: query
          description: true waits for the download to finish
          schema:
            type: string
            enum: ["true", "false"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InstallFromURLRequest"
      responses:
        "200":
          description: Started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  model_id:
                    type: string
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/models/{id}/install:
    post:
      summary: Download a catalog model, here or on a paired computer
      operationId: installModel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: wait
          in: query
          schema:
            type: string
            enum: ["true", "false"]
        - name: node_id
          in: query
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                node_id:
                  type: string
                wait:
                  type: boolean
      responses:
        "200":
          description: Started, or installed with wait
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [started, installed]
                  model_id:
                    type: string
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/models/{id}/start:
    post:
      summary: Load a model
      operationId: startModel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                node_id:
                  type: string
      responses:
        "200":
          description: The running instance
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/models/{id}/stop:
    post:
      summary: Unload a model instance
      operationId: stopModel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                node_id:
                  type: string
                instance_id:
                  type: string
                  description: Defaults to the path id
      responses:
        "200":
          description: Stopped
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/models/{id}:
    delete:
      summary: Delete a model file
      operationId: deleteModel
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: node_id
          in: query
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                node_id:
                  type: string
      responses:
        "204":
          description: Deleted
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/runtimes/{id}/install:
    post:
      summary: Install a runtime, such as llamacpp
      operationId: installRuntime
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Installed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  runtime_id:
                    type: string
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/profiles/{id}/reset:
    post:
      summary: Put a built-in profile back to how Toskar ships it
      operationId: resetProfile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AIProfile"
        "400":
          description: Not a built-in profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/chat/stop:
    post:
      summary: Stop a conversation's running turn
      operationId: stopChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                conversation_id:
                  type: string
              required: [conversation_id]
      responses:
        "200":
          description: Whether a turn was stopped
          content:
            application/json:
              schema:
                type: object
                properties:
                  stopped:
                    type: boolean
...
  /api/v1/tasks/{id}:
    get:
      summary: One task
      operationId: getTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: An API key from the API Access page, or TOSKAR_API_KEY
  schemas:
...
    AIProfile:
      type: object
      additionalProperties: true
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    InstallFromURLRequest:
      type: object
      properties:
        source_url:
          type: string
        display_name:
          type: string
        id:
          type: string
        filename:
          type: string
        size_bytes:
          type: integer
        parameters:
          type: string
        variant:
          type: string
        tags:
          type: array
          items:
            type: string
        node_id:
          type: string
      required: [source_url]
...